backend/pyproject.toml
[project]
name = "chronos-backend"
version = "0.1.0"
description = "Chronos - AI-powered incident management platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "redis[hiredis]>=5.0",
    "pyjwt>=2.9",
    "bcrypt>=4.0",
    "langgraph>=0.4",
    "langgraph-checkpoint-postgres>=2.0",
    "langchain-core>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.3",
    "cryptography>=44.0",
    "psycopg[binary]>=3.2",
    "pydantic-settings>=2.6",
    "loguru>=0.7",
    "orjson>=3.10",
    "httpx>=0.28",
    "python-multipart>=0.0.18",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "ruff>=0.9",
    "poethepoet>=0.32",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chronos"]

[tool.poe.tasks]
dev = "uvicorn chronos.main:app --reload --host 0.0.0.0 --port 8000"
migrate = "alembic upgrade head"
makemigrations = "alembic revision --autogenerate"
lint = "ruff check src/"
fix = "ruff check src/ --fix"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
